Add a de.json
I put together a translation - Line 205 "check" is that a verb, as in "check that this is right", or "check this box" or is it a noun "this is a check for agility"
I tried to trace where words would go in the layout, but couldn't for everything. If there are abbrev. needed for layout reasons, I am happy to help. ...
One question: Line 205 "check" is that a verb, as in "check that this is right", or "check this box" or is it a noun "this is a check for agility" . Is it the tool tip for GM only?