fix(Manifest): system.json manifest address
This should fix the update problem we have when trying to update the game system in the Foundry client.

Explanation: When a new update is released, the Foundry client can see it, but it can install it. It's because the system.json file on our computer has a manifest address actually pointing to the older version.