armor mod dialog no longer prompts for stress.
rollAbilityMod now strictly gathers modifiers from the user and then calls rollAbility for the actual roll.  This results in reduction of duplicate programming logic from these 2 methods (rollAbility and rollAbilityMod).

dataset.mod and dataset.stressMod are set to 0 explicitly inside rollAbility because the DOM remembers these values and there are situations it can remember the previous rolls modifier if the user doesn't create another dialog.

There is still a bit more refactoring to do, but I want to check this version in because it is tested and works.