Synthetics now have a checkbox which sets whether or not they "immitate" human panic and stress rolls and render a push button. IF checked then the words "NO STRESS" on the character sheet are a button which will make a panic roll. The synth can make panic rolls if checked, but the call to rollAbility will  that never make rolls above 6.
This means pressing "NO STRESS" will never actually cause the synth to panic.

This checkbox will also cause a PUSH button to appear on applicable skill and ability checks, but it wont cause a synthetic to actually push.

Secret synthetic players should turn this checkbox on.  Open synth characters should turn this checkbox off.

Localization needs to be completed for languages other than english, or else the label will appear in english on the character sheet.

NO CSS was touched for this change.