make xeno rolls into ROLL chatmessages.  The only noticible difference here is that the chat message is considered a "roll" which means for for the Tabbed Chatlog page it appears in the correct page - (the roll page) rather than the out-of-character-page. This also means tools and scripts that filter on rolls will catch the rolls, and that when dicesonice is not used it should make the roll dice sound.
The Whisper setting is honoured which makes this into a private GM roll, and it appears as a blind GM roll in the roll page.    One might Argue the GM should be allow to roll the xeno attacks immediately as public. One may also argue it ought to just follow the Chat sidebar roll visibility settings exactly like other rolls. Maybe this can be a system setting.