setupCombatantCloning - modified to make the code changes take place prior to any modules. this will prevent libWrapper from complaining there there is a conflicting module with no name. Also this is necessary because a module that wraps Combat.prototype.createEmbeddedEntity with libwrapper will cause the cloning to break because rather than wrapping this version for the alienrpg system, libwrapper invokes the base system version, because the alienrpg version didnt get loaded until Hooks.once('ready')  which is very late.    This changes it so the alienrpg version is installed immedately after the foundry base system - which is most logical.  Foundry -> System -> Modules is the logical order for code to be installed.
The overwridden method for Combat.prototype.createEmbeddedEntity was not returning the Promise from the base call. This is a potential source of incompatability. This method is supposed to return a Promise. Fixed.

The overwridden system clones were using the 'hidden' attribute of the Token rather than the hidden attribute of the combatant.    This doesn't cause a problem per se but it can mean that the clone will be visible even if the original combatant is hidden.  The combatant 'hidden' status is what governs the visibility of combatants. the hidden status of a token governs its visibility on the canvas. These are 2 separate attibutes.  This is important for compatability with a module Your Tokens Visible, which can set hostile new combatants to default to hidden. Without this fix, a xeno would appear with the original combatant hidden, but all the clones will be visible.   The clones should just honour the combatant hidden trait as they too are all combatants - not separate tokens.